Add files via upload
Initial Commit.

SRCBB_Scraper.ipynb -> The notebook used to scrape the basketball team data from the SRCBB website
TournamentBracket.py -> The official bracket that was released as of 3/17/2024. The First Four games have not been played yet. A random equivalent team was chosen for those games.
ScheduleGraphViz.ipynb -> A simple graph network to visualize the teams which played each other during the season.
replacement_dict.py -> Some teams are not always referred to the same. "UConn" on one website might be referred to as "Connecticut" on another.
Model.ipynb -> The neural network that uses the data from the above files to make winning predictions.
Bracket Prediction.pdf -> The output of the neural network when entered into the ESPN Men's Tournament Challenge Bracket website.